+++ backend/CHANGELOG.md
@@ -6,3 +6,4 @@
 - Handling of block `1385825` on testnet. Which changes the NextPaydayTime by just 287 seconds.
 - Added query to fetch backend version.
 - Sorting on Nodes Page
+- Added Accounts Updated GraphQl subscription. Can be used to subscribe to updates of a specific Account.